doit: Split out cmd_p2a
